backend/pyproject.toml
[project]
name = "validation-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "asyncmy>=0.2.10",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "duckdb>=1.4.3",
    "fastapi>=0.124.4",
    "faststream[cli,confluent,rabbit,redis]>=0.6.4",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-sdk",
    "sqlalchemy-utils>=0.42.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlglot>=28.3.0",
    "sqlmodel>=0.0.27",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
strict = true
pretty = true
show_error_codes = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
implicit_optional = false
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
show_error_code_links = true
color_output = true
show_absolute_path = true
plugins = ["returns.contrib.mypy.returns_plugin"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "build/",
    "dist/",
    "venv/",
    "\\.venv/",
    "tests/helpers/"
]


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
indent-width = 4



[tool.ruff.lint]
allowed-confusables = ["−", "ρ", "∗"]
fixable = ["ALL"]
ignore = []
select = ["E4", "E7", "E9", "F", "B", "Q"]
unfixable = []



[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401", "E402", "F403"]
"**/{tests,docs,tools}/*" = ["E402", "F403"]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.pytest.ini_options]
addopts = "--cov=src--cov-report=term-missing"



[tool.ty.rules]
unresolved-reference = "ignore"
unresolved-attribute = "ignore"
possibly-unbound-attribute = "ignore"
invalid-return-type = "ignore"
invalid-assignment = "ignore"
unresolved-import = "ignore"
invalid-argument-type = "ignore"
invalid-type-form = "ignore"
no-matching-overload = "ignore"
missing-argument = "ignore"

[tool.uv.sources]
python-sdk = { git = "https://github.com/shakedguy/python-sdk.git" }

[dependency-groups]
dev = [
    "icecream>=2.1.8",
    "ipython>=9.8.0",
    "mypy>=1.19.0",
    "pandas-stubs>=2.3.3.251201",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]


[tool.alembic]

# path to migration scripts.
# this is typically a path given in POSIX (e.g. forward slashes)
# format, relative to the token %(here)s which refers to the location of this
# ini file
script_location = "%(here)s/alembic"

# template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
# Uncomment the line below if you want the files to be prepended with date and time
# see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
# for all available tokens
# file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s"

# additional paths to be prepended to sys.path. defaults to the current working directory.
prepend_sys_path = [
    "."
]

# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the tzdata library which can be installed by adding
# `alembic[tz]` to the pip requirements.
# string value is passed to ZoneInfo()
# leave blank for localtime
# timezone =

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
# output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
# name = "ruff"
# type = "module"
# module = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
# name = "ruff"
# type = "exec"
# executable = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
